+++ packages/edit-site/src/components/page-patterns/index.js
@@ -34,6 +34,7 @@ import { usePrevious } from '@wordpress/compose';
 /**
  * Internal dependencies
  */
+import { Async } from '../async';
 import Page from '../page';
 import {
 	LAYOUT_GRID,
@@ -177,10 +178,12 @@ function Preview( { item, categoryId, viewType } ) {
 					{ isEmpty && isTemplatePart && __( 'Empty template part' ) }
 					{ isEmpty && ! isTemplatePart && __( 'Empty pattern' ) }
 					{ ! isEmpty && (
-						<BlockPreview
-							blocks={ item.blocks }
-							viewportWidth={ item.viewportWidth }
-						/>
+						<Async>
+							<BlockPreview
+								blocks={ item.blocks }
+								viewportWidth={ item.viewportWidth }
+							/>
+						</Async>
 					) }
 				</PreviewWrapper>
 			</div>
